requirements.txt
discord.py
python-dotenv
aiohttp
git+https://github.com/Pycord-Development/pycord
asyncio
discord_together